refactor(ui): split analysis panel sections
Extract focused analysis panel subcomponents and shared field/spec contracts so the main analysis panel file keeps only the bridge shell and local signal-driven effects.
